refactor(welcome): Placed sections of the Welcome.vue page in the correct folder.
