+++ resources/js/components/welcome/SectionConditionals.vue
