+++ resources/js/components/welcome/SectionControls.vue
