+++ resources/js/components/welcome/SectionCta.vue
