+++ resources/js/components/welcome/SectionFooter.vue
