+++ resources/js/components/welcome/SectionIntegrations.vue
