+++ resources/js/components/welcome/SectionOnboarding.vue
